Fix github workflow badge url
For some reason the SVG status badge URL in this repo does not work out-of-the-box
if copied from the 'Create Status Badge' dialog. Instead, the badge must
be referenced by the workflow 'name' (as specified in in go.yml), in this
case by the name 'Go'.

However on repo 'MawKKe/stopwatch-go' the badge URL works just as you
would expect without any modifications. What gives, Github???